[AD-522] save SSH values for re-read, make settings window into 2 columns
Make ssl and additional settings groups to be on the right side of the column on the configuration window. This is subject to change later.
I created the function
void DsnConfigurationWindow::RetrieveSshParameters(config::Configuration& cfg) const
to save SSH values.
The width value in the DsnConfigurationWindow constructor is doubled from 360 to 720 and added the margin value. The margins currently look imbalanced on the settings box, and will be fixed in later commits.
Make Ok button and Cancel buttons aligned with the right column
bugfix - allow the SSH strict host checking checkbox to be checked on and off by adding the case ChildId::SSH_STRICT_HOST_KEY_CHECKING_CHECK_BOX definition.