[AD-543] adapt ColumnsMetaQuery to accept catalog as a parameter
Functions affected:
* ColumnMetadataQuery
* ExecuteGetColumnsMetaQuery
* InternalExecuteGetColumnsMetaQuery
Reason for change: since the JNI call to getColumns accepts catalog as a parameter, it is required for those 3 functions to essentially pass catalog as a parameter as well, even though the JDBC driver doesn't actually deal with the value of catalog. JNI call is to be implemented in later commits.