[AD-543] remove JNI wrapper code for getShort and its unit test
reason: getShort JNI wrapper method call was implemented with the intent to use it for column metadata value SOURCE_DATA_TYPE (for getting the column metadata from the resultSet returned by getColumns), which is a short. However, ODBC driver does not need SOURCE_DATA_TYPE, thus getShort is of no use to JDBC at this point and should be removed.