[AD-771] fix issue of JDBC not returning tables
* Tableau passes catalog, schema, table name, and table type as null pointers to SQLTables.
PowerBI passes catalog, schema, and table name as null pointers to SQLTables.

PowerBI passes column name as null pointers to SQLColumns.

Therefore, if null is passed in ODBC as table pattern, "%" needs to be passed as table pattern to JDBC.
Same thing with null being passed to ODBC as column name pattern.